+++ engine/include/richard.h
@@ -1,5 +1,5 @@
 /*
-* This header contains all the headers that will be included in RichardEditor
+* This header contains all the headers that will be included in application
 */
 
 #include "core/engine.h"
