+++ engine/src/core/window.h
@@ -5,7 +5,7 @@
 * will not have to import anything SDL_Window related.
 * 
 * For example, engine.h will not need to import this struct
-* and neither will RichardEditor/src/main.cpp, as it imports engine.h
+* and neither will application/src/main.cpp, as it imports engine.h
 */
 struct SDL_Window;
